refactor: extract shared MIME detection utility
- Create utils/imageDownload.ts for reusable base64 conversion
- Apply to seedream, seedance, and nanobanana models
- Eliminates code duplication across 3 models
